ref: break server.js into a bootstrap
