+++ backend/server.js
@@ -460,114 +460,6 @@ process.on('SIGINT', async () => {
     }
 });
 
-// The subscription renewal sweep used to be scheduled here, as a bare
-// top-level `setInterval(processRenewals, 24 * 60 * 60 * 1000)` (#1494). It is
-// scheduled with the other three background jobs in bootstrap() now, for the
-// three reasons it should always have been:
-//
-//   * it sat outside the `NODE_ENV !== "test"` guard the others are behind and
-//     had no unref, so it armed itself during `npm test` and held the event
-//     loop open -- part of why the suite ends with "Force exiting Jest";
-//   * the first run was 24 hours after boot and the timer reset on every
-//     restart, so a service deploying more than once a day never swept;
-//   * it had no try/catch of its own, unlike the three below it.
-
-// 11. Application Bootstrap Function
-async function bootstrap() {
-    const { logServerStartup } = require('./utils/serverStartupLogger');
-    console.log("Initializing core background services...");
-
-    const services = [
-        { name: 'HealthScoreService', instance: healthScoreService },
-        { name: 'MetricsAggregationService', instance: metricsAggregationService },
-        { name: 'TracingService', instance: tracingService },
-        { name: 'PolicyEngineService', instance: policyEngine },
-        { name: 'OutboxService', instance: outboxService },
-        { name: 'FeatureFlagService', instance: featureFlagService },
-        { name: 'SLAService', instance: slaService },
-        { name: 'LoyaltyService', instance: loyaltyService },
-        { name: 'ProvenanceService', instance: provenanceService },
-        { name: 'CapabilityMappingService', instance: capabilityMappingService },
-        { name: 'PluginSystem', instance: pluginSystem },
-        { name: 'JobQueue', instance: jobQueue }
-    ];
-
-    for (const s of services) {
-        try {
-            await s.instance.initialize();
-            console.log(`Service '${s.name}' initialized successfully.`);
-        } catch (err) {
-            console.error(`Warning: Service '${s.name}' failed to initialize:`, err.message);
-        }
-    }
-
-    try {
-        initializeContainer();
-        console.log("DI Container initialized successfully.");
-    } catch (err) {
-        console.error("Warning: DI Container initialization failed:", err.message);
-    }
-
-    try {
-        setupAllSubscribers();
-        console.log("Event subscribers set up successfully.");
-    } catch (err) {
-        console.error("Warning: Failed to setup event subscribers:", err.message);
-    }
-
-    // Periodic back-in-stock / price-drop scan (#1233). No-ops under test.
-    if (process.env.NODE_ENV !== "test") {
-        try {
-            startStockAlertScheduler();
-        } catch (err) {
-            console.error("Warning: Failed to start stock-alert scheduler:", err.message);
-        }
-
-        try {
-            const { startPriceDropJob } = require("./jobs/priceDropJob");
-            startPriceDropJob();
-        } catch (err) {
-            console.error("Warning: Failed to start wishlist price-drop job:", err.message);
-        }
-
-        // Abandoned-cart recovery (#1429). The lifecycle sweep decides what is
-        // abandoned; this decides what to say about it, and to whom.
-        try {
-            const { startCartRecoveryJob } = require("./jobs/cartRecoveryJob");
-            startCartRecoveryJob();
-        } catch (err) {
-            console.error("Warning: Failed to start cart recovery job:", err.message);
-        }
-
-        // Subscription renewals (#1494). Moved here from a bare top-level
-        // setInterval so it is guarded, caught and stoppable like the rest.
-        try {
-            const {
-                startSubscriptionRenewalJob
-            } = require("./jobs/subscriptionRenewalJob");
-            startSubscriptionRenewalJob();
-        } catch (err) {
-            console.error("Warning: Failed to start subscription renewal job:", err.message);
-        }
-    }
-
-    // Start HTTP listening only after services finish initializations
-    console.log("Starting HTTP server...");
-    server.listen(PORT, "0.0.0.0", () => {
-        logServerStartup({
-            port: PORT,
-            environment: process.env.NODE_ENV || "development",
-            frontendUrl: FRONTEND_URL,
-            logsDir: logDir,
-            healthUrl: `http://localhost:${PORT}/health`,
-            mcpSecurity: true,
-            rateLimiting: true,
-            helmet: true,
-        });
-    });
-}
-
-// Start application
-bootstrap();
+app.server = server;
 
 module.exports = app;
